Do plugin configuration in a plugin section in gerrit.config
Plugins should not introduce new sections in the gerrit.config file,
but define their configuration parameters in a plugin subsection.

So instead of having

  [avatar]
    url = ...

the configuration must now be

  [plugin "<plugin-name>"]
    url = ...

Since we just created the project newly to fix the project name, it's
a good opportunity to change the configuration as well.

Change-Id: Icba8d50185bd542e5a3900e8a19b48284ba46212
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>