Update patch set 1
Patch Set 1:

Hey Saša, thanks a lot for reviewing and for your patience: I don't use Java very often, so there are lots of pitfalls for me to run into.

The Gerrit contributor guidelines don't say anything about Verified+1 for own contributions: I verified that this works, so can I V+1 myself or does somebody else need to do it?

It's running in production on review.coreboot.org, with just a minor patch to deal with differences to the stable plugin API to use getPreferredEmail() instead of preferredEmail().

Patch-set: 1